v16.4.2 build 3 NOTES - Fixed: Shop items showing "E7", "E3" etc after their price if above a certain amount

+++ randompackage-spigot/src/main/java/me/randomhashtags/randompackage/addons/objects/ShopItem.java
@@ -3,12 +3,14 @@ package me.randomhashtags.randompackage.addons.objects;
 import me.randomhashtags.randompackage.utils.universal.UMaterial;
 import org.bukkit.inventory.ItemStack;
 
+import java.math.BigDecimal;
+
 public class ShopItem {
     public final String path, opensCategory;
     public final int slot;
     private final ItemStack display, purchasedItem;
-    public final double buyPrice, sellPrice;
-    public ShopItem(String path, int slot, String opensCategory, ItemStack display, ItemStack purchasedItem, double buyPrice, double sellPrice) {
+    public final BigDecimal buyPrice, sellPrice;
+    public ShopItem(String path, int slot, String opensCategory, ItemStack display, ItemStack purchasedItem, BigDecimal buyPrice, BigDecimal sellPrice) {
         this.path = path;
         this.slot = slot;
         this.opensCategory = opensCategory;
