+++ randompackage-spigot/src/main/java/me/randomhashtags/randompackage/api/events/ShopPrePurchaseEvent.java
@@ -4,13 +4,15 @@ import me.randomhashtags.randompackage.addons.objects.ShopItem;
 import org.bukkit.entity.Player;
 import org.bukkit.event.Cancellable;
 
+import java.math.BigDecimal;
+
 public class ShopPrePurchaseEvent extends AbstractEvent implements Cancellable {
     private boolean cancelled;
     public final Player player;
     public final ShopItem item;
     private int amount;
-    private double cost;
-    public ShopPrePurchaseEvent(Player player, ShopItem item, int amount, double cost) {
+    private BigDecimal cost;
+    public ShopPrePurchaseEvent(Player player, ShopItem item, int amount, BigDecimal cost) {
         this.player = player;
         this.item = item;
         this.amount = amount;
@@ -18,8 +20,8 @@ public class ShopPrePurchaseEvent extends AbstractEvent implements Cancellable {
     }
     public int getAmount() { return amount; }
     public void setAmount(int amount) { this.amount = amount; }
-    public double getCost() { return cost; }
-    public void setCost(double cost) { this.cost = cost; }
+    public BigDecimal getCost() { return cost; }
+    public void setCost(BigDecimal cost) { this.cost = cost; }
 
     public boolean isCancelled() { return cancelled; }
     public void setCancelled(boolean cancel) { cancelled = cancel; }
