+++ randompackage-spigot/src/main/java/me/randomhashtags/randompackage/api/events/ShopSellEvent.java
@@ -4,13 +4,15 @@ import me.randomhashtags.randompackage.addons.objects.ShopItem;
 import org.bukkit.entity.Player;
 import org.bukkit.inventory.ItemStack;
 
+import java.math.BigDecimal;
+
 public class ShopSellEvent extends AbstractEvent {
     public final Player player;
     public final ShopItem shopitem;
     public final ItemStack item;
     public final int amount;
-    public final double profit;
-    public ShopSellEvent(Player player, ShopItem shopitem, ItemStack item, int amount, double profit) {
+    public final BigDecimal profit;
+    public ShopSellEvent(Player player, ShopItem shopitem, ItemStack item, int amount, BigDecimal profit) {
         this.player = player;
         this.shopitem = shopitem;
         this.item = item;
